requirements.txt
fastapi
uvicorn[standard]
psycopg[binary]
bcrypt
pydantic
pydantic-settings
python-dotenv
pyjwt
